refactor v2 trace pipe steps
